[Console/Input] input\parameter\Values: Moved the logic responsible for populating Values collections with defaults out of all() into a public finalize() method, to make this a one-off call during input binding.
As a by-product, get() no longer attempts to resolve a Parameter's default value if the value is not explicitly set - it now instead silently assumes the collection has been finalized and includes default values.

Immutability, however, is *not* enforced. Values can currently be freely modified/unset even after finalizing the collection. While convenient, this can lead to inconsistent data, and may therefore be changed in the future.